Feature: SET_AP_TO_PREV_BLOCK_HASH and SET_AP_TO_NEW_BLOCK_HASH hints
